build(wasm): unbreak emcc build, add `--wasm` to dev sweep
WASM build was broken on main before today (5 conflicting-types errors
under emcc's -Werror=implicit-function-declaration; the 2026-05-14
smaller-items sweep also introduced shift-count-overflow warnings on
wasm32 via JACL_PACK_PTR).

- Forward-declare vm__fmt_value (vm.c) and the four
  runtime__init_state / start_threads / stop_threads / teardown_state
  functions (runtime.c) so the unity build's first-use sites have a
  prototype instead of an implicit declaration.
- JACL_PACK_PTR: cast uintptr_t → uint64_t *before* the >> 56 in the
  debug assert. uintptr_t is 32 bits on wasm32; shifting it by 56 is
  UB. Behavior on 64-bit hosts is unchanged. Fixed in both jacl.h and
  value.c's local copy.
- build.sh --wasm: new flag, compile-only check via emcc. Skips
  cleanly with a printed notice when emcc isn't on PATH (so it stays
  out of devs' way) and rejects --wasm + --tsan as mutually
  exclusive. AUDIT.md test-baselines section names it as a third
  pre-merge check alongside normal-mode and TSAN.

WASM build is now clean (17 pre-existing tautological-comparison
warnings remain; none are real). Normal mode 88/88.

+++ src/jacl.h
@@ -47,9 +47,12 @@ typedef uint64_t JaclVal;
  * catches a pointer wider than the 56-bit payload — without it, the
  * mask would silently truncate high bits into the tag. The mask itself
  * stays for release-build defense; with the assert holding, it's a
- * no-op on every supported platform today (48-bit user-space VA). */
+ * no-op on every supported platform today (48-bit user-space VA).
+ *
+ * Note: cast to uint64_t *before* the shift — on wasm32 / other 32-bit
+ * targets, uintptr_t is 32 bits, so a >> 56 on uintptr_t is UB. */
 #define JACL_PACK_PTR(tag, p)                                                 \
-    (assert(((uintptr_t)(p) >> JACL_TAG_SHIFT) == 0),                         \
+    (assert(((uint64_t)(uintptr_t)(p) >> JACL_TAG_SHIFT) == 0),               \
      (tag) | ((uint64_t)(uintptr_t)(p) & JACL_PAYLOAD_MASK))
 
 #define JACL_FLAG_TAINTED   ((uint64_t)1 << 63)
